+++ examples/freertos/compact_log.h
@@ -1,0 +1,16 @@
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void compact_log_cpp_example(void);
+
+#ifdef __cplusplus
+}
+#endif