+++ ports/zephyr/v2.4/memfault_fault_handler.c
@@ -24,13 +24,10 @@
 // Starting in v3.4, the handler set function was renamed and the declaration
 // added to a public header
 #if MEMFAULT_ZEPHYR_VERSION_GT(3, 3)
-  // The nmi.h header was moved after Zephyr v3.4. Support both locations.
-  #if __has_include(MEMFAULT_ZEPHYR_INCLUDE(arch/arm/aarch32/nmi.h))
-    #include MEMFAULT_ZEPHYR_INCLUDE(arch/arm/aarch32/nmi.h)
-  #elif __has_include(MEMFAULT_ZEPHYR_INCLUDE(arch/arm/nmi.h))
-    #include MEMFAULT_ZEPHYR_INCLUDE(arch/arm/nmi.h)
+  #if MEMFAULT_ZEPHYR_VERSION_GT(3, 4)
+    #include <cmsis_core.h>
   #else
-    #error "Missing Zephyr nmi.h header"
+    #include MEMFAULT_ZEPHYR_INCLUDE(arch/arm/aarch32/nmi.h)
   #endif
   #define MEMFAULT_ZEPHYR_NMI_HANDLER_SET z_arm_nmi_set_handler
 #else